Make border to Settings search transparent when selected
We want the border to blend with the background no matter what
style theme we use - this way we don't end up with a thick white
bix around the search fieldon dark themes (or vice versa on light
themes). So why not just make it transparent to resolve this? :P

Change-Id: Ia52c681cf1deba2d68596f0282c6c0bd77113a41
Signed-off-by: calebcabob <calphonic@gmail.com>